TW-1088 Update document
