+++ example/lib/data/fcm_notification.dart
@@ -9,12 +9,16 @@ class FcmNotification {
 
   static const keyTransactionHash = 'tnx_hash';
   static const keyStatus = 'status';
+  static const keyOrderID = 'order_id';
 
   FcmNotification.fromMap(Map<String, dynamic> map) {
-    TChainPaymentStatus status = map[keyStatus] == 2
+    TChainPaymentStatus status = map[keyStatus] == '2'
         ? TChainPaymentStatus.success
         : TChainPaymentStatus.proceeding;
     result = TChainPaymentResult(
-        status: status, transactionID: map[keyTransactionHash]);
+      status: status,
+      orderID: map[keyOrderID] as String? ?? '',
+      transactionID: map[keyTransactionHash] as String?,
+    );
   }
 }
